Revert "SEO optimization"
