+++ src/app/layout.js
@@ -1,13 +1,25 @@
-import { Metadata } from 'next'
+import localFont from "next/font/local";
+import "./globals.css";
+
+const geistSans = localFont({
+  src: "./fonts/GeistVF.woff",
+  variable: "--font-geist-sans",
+  weight: "100 900",
+});
+const geistMono = localFont({
+  src: "./fonts/GeistMonoVF.woff",
+  variable: "--font-geist-mono",
+  weight: "100 900",
+});
 
 export const metadata = {
-  metadataBase: new URL(process.env.NEXT_PUBLIC_SITE_URL || 'http://localhost:3000'),
+  metadataBase: new URL(process.env.NEXT_PUBLIC_SITE_URL || 'https://asian-scales.vercel.app'),
   title: {
-    default: 'Asian Scales - Industrial Weighing Solutions',
+    default: 'Asian Scales - Weighing & Automation Solutions',
     template: '%s | Asian Scales'
   },
-  description: 'Leading provider of industrial weighing solutions, automation systems, and trading services in Asia',
-  keywords: ['industrial scales', 'weighing solutions', 'automation', 'trading', 'Asian Scales'],
+  description: 'Leading provider of industrial weighing scales, automation solutions, and trading services in Asia. Discover our range of precision measurement equipment.',
+  keywords: ['weighing scales', 'industrial automation', 'measurement equipment', 'Asian Scales', 'trading services'],
   robots: {
     index: true,
     follow: true,
@@ -21,28 +33,17 @@ export const metadata = {
   },
   verification: {
     google: 'your-google-verification-code',
-  },
-  openGraph: {
-    title: 'Asian Scales - Industrial Weighing Solutions',
-    description: 'Leading provider of industrial weighing solutions, automation systems, and trading services in Asia',
-    url: process.env.NEXT_PUBLIC_SITE_URL,
-    siteName: 'Asian Scales',
-    locale: 'en_US',
-    type: 'website',
-  },
-  twitter: {
-    card: 'summary_large_image',
-    title: 'Asian Scales - Industrial Weighing Solutions',
-    description: 'Leading provider of industrial weighing solutions, automation systems, and trading services in Asia',
   }
 }
 
-export default function RootLayout({
-  children
-}) {
+export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body
+        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+      >
+        {children}
+      </body>
     </html>
-  )
+  );
 }
